chore: donation page grammar fix
Added commas for better flow. Added hyphen. Use "is" instead of "are" when the object of the sentence is not plural. If I may make a suggestion, the Verge team needs to pay closer attention to things like this. Especially on the official Twitter, which freqently has missing capital letters, lack of punctuation, typos, etc.